Moving MapZipper test to correct location
[ghstack-poisoned]
